Sms service integration rel dev

+++ routes/api.php
@@ -20,10 +20,10 @@ Route::group([
     'middleware' => 'api',
     'prefix' => 'auth'
 ], static function () {
-    Route::post('login', [AuthController::class,'login']);
-    Route::get('logout',  [AuthController::class,'logout']);
-    Route::get('refresh',  [AuthController::class,'refresh']);
-    Route::get('me',  [AuthController::class,'me']);
+    Route::post('login', [AuthController::class, 'login']);
+    Route::get('logout', [AuthController::class, 'logout']);
+    Route::get('refresh', [AuthController::class, 'refresh']);
+    Route::get('me', [AuthController::class, 'me']);
 });
 
 Route::prefix('user')->group(
@@ -31,11 +31,9 @@ Route::prefix('user')->group(
         Route::get('/{user}', [UserController::class, 'show']);
         Route::get('/', [UserController::class, 'index']);
         Route::post('/', [UserController::class, 'create']);
-        Route::post('/profile/{user}', [UserProfileController::class, 'fillProfile'])->name('profile.update');
-        Route::patch('/{user}', [UserController::class, 'update']);
-        Route::delete('/{user}', [UserController::class, 'delete'])->middleware('admin');
+        Route::post('/profile/{user}', [UserProfileController::class, 'fillProfile'])
+            ->middleware('auth')
+            ->name('profile.update');
+        Route::patch('/{user}', [UserController::class, 'update'])->middleware('auth');
+        Route::delete('/{user}', [UserController::class, 'delete'])->middleware(['auth', 'admin']);
     });
-
-//Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
-//    return $request->user();
-//});
